Extract KtConstructor creator
